feat: add book to cart

+++ internal/model/order_model.go
@@ -5,30 +5,30 @@ import (
 )
 
 type Order struct {
-	ID         int       `json:"id"`
-	CustomerID int       `json:"customer_id"`
+	ID         int64     `json:"id"`
+	CustomerID int64     `json:"customer_id"`
 	UpdatedAt  time.Time `json:"updated_at"`
-	OrderState int       `json:"order_state"`
+	OrderState int64     `json:"order_state"`
 	Total      float64   `json:"total"`
 }
 
 type OrderDetail struct {
-	ID       int     `json:"id"`
-	OrderID  int     `json:"order_id"`
-	BookID   int     `json:"book_id"`
-	Quantity int     `json:"quantity"`
+	ID       int64   `json:"id"`
+	OrderID  int64   `json:"order_id"`
+	BookID   int64   `json:"book_id"`
+	Quantity int64   `json:"quantity"`
 	Subtotal float64 `json:"subtotal"`
 }
 
 type OrderResponse struct {
-	ID          int                   `json:"id"`
+	ID          int64                 `json:"id"`
 	OrderDetail []OrderDetailResponse `json:"orderDetails"`
 	Total       float64               `json:"total"`
 }
 
 type OrderDetailResponse struct {
-	ID       int     `json:"id"`
+	ID       int64   `json:"id"`
 	Book     []Book  `json:"books"`
-	Quantity int     `json:"quantity"`
+	Quantity int64   `json:"quantity"`
 	Subtotal float64 `json:"subtotal"`
 }
